React Native upgrade for rutime behavior
