+++ package.json
@@ -74,21 +74,23 @@
     "moment-timezone": "^0.5.17",
     "prop-types": "^15.5.10",
     "query-string": "^4.1.0",
-    "react": "^16.6.0",
+    "react": "16.6.0-alpha.8af6728",
     "react-native": "0.57.4",
     "react-native-hyperlink": "0.0.11",
     "react-native-parallax-scroll-view": "https://github.com/orta/react-native-parallax-scroll-view",
-    "react-native-scrollable-tab-view": "^0.8.0",
+    "react-native-scrollable-tab-view": "^0.9.0",
     "react-native-sentry": "^0.30.2",
     "react-relay": "https://github.com/alloy/relay/releases/download/v1.5.0-artsy.5/react-relay-1.5.0-artsy.5.tgz",
     "react-tracking": "^5.0.0",
     "relay-runtime": "https://github.com/alloy/relay/releases/download/v1.5.0-artsy.5/relay-runtime-1.5.0-artsy.5.tgz",
     "remove-markdown": "0.1.0",
     "simple-markdown": "^0.4.0",
-    "styled-components": "3.2.2",
+    "styled-components": "3.4.9",
     "tipsi-stripe": "5.2.4"
   },
   "devDependencies": {
+    "@babel/core": "^7.0.0",
+    "@babel/plugin-proposal-decorators": "^7.1.2",
     "@playlyfe/gql": "^2.3.2",
     "@storybook/react-native": "^3.3.11",
     "@types/jest": "^22.1.1",
@@ -103,30 +105,25 @@
     "@types/relay-runtime": "^1.3.5",
     "@types/styled-system": "^3.0.1",
     "awesome-typescript-loader": "^3.2.3",
-    "babel-jest": "^22.1.0",
+    "babel-jest": "23.6.0",
     "babel-plugin-relay": "https://github.com/alloy/relay/releases/download/v1.5.0-artsy.3/babel-plugin-relay-1.5.0-artsy.3.tgz",
-    "babel-polyfill": "^6.13.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-react-native": "^4.0.0",
-    "babel-preset-stage-0": "^6.24.1",
     "concurrently": "^2.2.0",
     "danger": "^3",
     "danger-plugin-jest": "^1.1.0",
     "enzyme": "^3.1.0",
     "enzyme-adapter-react-16": "^1.0.2",
     "husky": "^0.14.3",
-    "jest": "^22.4.2",
+    "jest": "23.6.0",
     "jest-styled-components": "^4.10.0",
     "lint-staged": "^6.0.0",
+    "metro-react-native-babel-preset": "0.49.0",
     "patch-package": "^5.0.0",
     "postinstall-prepare": "^1.0.1",
     "prettier": "^1.14",
     "react-dom": "^16.3.0-alpha.1",
     "react-native-storybook-loader": "^1.6.0",
-    "react-native-typescript-transformer": "^1.2.3",
     "react-storybooks-relay-container": "^1.0.0",
-    "react-test-renderer": "^16.3.0-alpha.1",
+    "react-test-renderer": "16.6.0-alpha.8af6728",
     "recursive-readdir-sync": "^1.0.6",
     "relay-compiler": "https://github.com/alloy/relay/releases/download/v1.5.0-artsy.5/relay-compiler-1.5.0-artsy.5.tgz",
     "relay-compiler-language-typescript": "^1.1.0",
